requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.1.0
anthropic>=0.40.0
